wispr-backend/requirements.txt
# Django & DRF
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# WebSocket & Async
channels==4.0.0
channels-redis==4.1.0
websockets==12.0

# JWT Auth
djangorestframework-simplejwt==5.3.2

# API Clients
requests==2.31.0
axios

# Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Additional utilities
python-decouple==3.11.8

daphne

google-generativeai==0.8.6

requests
websockets

python-dotenv
 
uvicorn

web: uvicorn config.asgi:application --host 0.0.0.0 --port 10000